+++ model/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @platforma-open/milaboratories.humanization-score.model
+# @platforma-open/milaboratories.humanness-score.model
 
 ## 0.3.2
 
@@ -24,7 +24,7 @@
 
 ### Minor Changes
 
-- 5c4b4f0: Initial release of the Humanization Score block.
+- 5c4b4f0: Initial release of the Humanness Score block.
 
   Scores antibody sequences for humanness relative to natural human antibody
   repertoires using the offline OASis/SwissProt databases bundled with the